Time: 4 ms (34.89%), Space: 44.1 MB (45.23%) - LeetHub

+++ 0496-next-greater-element-i/0496-next-greater-element-i.java
@@ -24,7 +24,7 @@ class Solution {
        int i=0;
        for(int val:nums1)
        {
-         result[i++]=(int)map.get(val);
+         result[i++]=map.get(val);
        }
        return result;
